#10 Missing VRAM constraints coefficients are added. Each layer allocated to GPU (VRAM for CUDA, RAM for unified memory devices) are now multiplied by bprime.
Additional minor adjustments in the comments.

All the parameters in the capacity constraints are converted to float for consistency as only some of them were converted to float.

Finally, a minor adjustment in the objective function coefficients to separate penalization of the devices in sets M2 (Mac w Metal) and M3 (Linux/Android) for GPU allocated layers, in accordance with the HALDA paper.