A new concept of cycle time to mimic overlap of the disk prefetch on a device, and compute on other devices, is introduced. This way, when computing the total latency, we can consider disk load time of the prefetch for all rounds (except the first one). Then, the cycle time, C, is multiplied by k-1 in the objective. Furthermore, for the first round as all the layers are assumed to be already allocated in the RAM/VRAM, the previous objective is kept for only one round, without any multiplication by k.
Busy_row calculates the total compute time and inter device comm time including the disk read time for only a few layers if necessary during a window is being computed, for a device.

Fetch_row computes the total disk read time for fetching the layers for the upcoming round.

Then, for each device i, we have two new constraints:

(1) C >= B_i + z_i

and

(2) z_i >= F_i - (C - B_i)

The first constraints ensures that the Cycle time for all devices is larger than or equal to each device's compute time plus the waiting time, where the waiting time is, as defined in the second constraint, larger than or equal to the difference between fetching time and the idle time (difference between cycle time and compute time), for each device. Therefore, if any latency is caused by the prefetch at a device, other devices can benefit from it as a buffer for their prefetch and compute. However, at the same time, the additional latency will naturally result in an increase in the total latency.